Rename Student class filename from person.ts to student.ts
