+++ src/app/dashboard/student-preview/student-preview.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { Student } from '../../shared/models/person';
+import { Student } from '../../shared/models/student';
 import { Colors } from '../../shared/constants/color.constants';
 import { Skill, SkillLevel } from '../../shared/models/skill';
 import { Device } from '../../shared/models/device';
