+++ src/app/shared/layers/business-logic-layer/team-generation/team-generation.service.ts
@@ -1,4 +1,4 @@
-import { Student } from '../../../models/person';
+import { Student } from '../../../models/student';
 import { Team } from '../../../models/team';
 /**
  * Created by Malte Bucksch on 25/11/2016.
