+++ src/app/shared/models/person-constraints/person-constraint.ts
@@ -1,4 +1,4 @@
-import { Student } from '../person';
+import { Student } from '../student';
 
 export abstract class PersonConstraint {
   abstract isFullfilledFor(person: Student): boolean;
